+++ apps/cli/src/commands/pact/index.ts
@@ -7,6 +7,9 @@ export const pactCommand = defineCommand({
   },
   subCommands: {
     install: async () => (await import('./install')).installCommand,
+    list: async () => (await import('./list')).listCommand,
     upgrade: async () => (await import('./upgrade')).upgradeCommand,
+    use: async () => (await import('./use')).useCommand,
+    uninstall: async () => (await import('./uninstall')).uninstallCommand,
   },
 });