+++ examples/todo-mvc-common/src/api/api3.spec.ts
@@ -1,6 +1,6 @@
 import { createPactTestEnv, getUuid } from 'pact-toolbox';
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
-import { createTodo, readTodos } from './api';
+import { createTodo, deleteTodoById, editTodoById, readTodoById, readTodos, toggleTodoStatusById } from './api';
 
 describe('todos api pact local server', async () => {
   let id = getUuid();
@@ -31,30 +31,30 @@ describe('todos api pact local server', async () => {
     expect(todos[0].title).toEqual(title);
   });
 
-  // it('should read one todo by id', async () => {
-  //   const todo = await readTodoById(id);
-  //   expect(todo.title).toEqual(title);
-  // });
+  it('should read one todo by id', async () => {
+    const todo = await readTodoById(id);
+    expect(todo.title).toEqual(title);
+  });
 
-  // it('should edit a todo', async () => {
-  //   const newTitle = 'build a todo app';
-  //   const message = await editTodoById(id, newTitle);
-  //   expect(message).toEqual('Write succeeded');
-  //   const t = await readTodoById(id);
-  //   expect(t.title).toEqual(newTitle);
-  // });
+  it('should edit a todo', async () => {
+    const newTitle = 'build a todo app';
+    const message = await editTodoById(id, newTitle);
+    expect(message).toEqual('Write succeeded');
+    const t = await readTodoById(id);
+    expect(t.title).toEqual(newTitle);
+  });
 
-  // it('should complete a todo', async () => {
-  //   const message = await toggleTodoStatusById(id);
-  //   expect(message).toEqual('Write succeeded');
-  //   const t = await readTodoById(id);
-  //   expect(t.completed).toEqual(true);
-  // });
+  it('should complete a todo', async () => {
+    const message = await toggleTodoStatusById(id);
+    expect(message).toEqual('Write succeeded');
+    const t = await readTodoById(id);
+    expect(t.completed).toEqual(true);
+  });
 
-  // it('should delete a todo', async () => {
-  //   const message = await deleteTodoById(id);
-  //   expect(message).toEqual('Write succeeded');
-  //   const t = await readTodoById(id);
-  //   expect(t.deleted).toEqual(true);
-  // });
+  it('should delete a todo', async () => {
+    const message = await deleteTodoById(id);
+    expect(message).toEqual('Write succeeded');
+    const t = await readTodoById(id);
+    expect(t.deleted).toEqual(true);
+  });
 });