+++ examples/todo-mvc-next/CHANGELOG.md
@@ -1,5 +1,14 @@
 # todo-mvc-next
 
+## 0.1.3
+
+### Patch Changes
+
+- more pact version manager features
+- Updated dependencies
+  - todo-mvc-common@0.0.3
+  - @pact-toolbox/client-utils@0.0.4
+
 ## 0.1.2
 
 ### Patch Changes