+++ packages/client-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pact-toolbox/client-utils
 
+## 0.0.4
+
+### Patch Changes
+
+- more pact version manager features
+
 ## 0.0.3
 
 ### Patch Changes