+++ packages/init/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pact-toolbox/init
 
+## 0.0.5
+
+### Patch Changes
+
+- more pact version manager features
+- Updated dependencies
+  - @pact-toolbox/utils@0.0.4
+
 ## 0.0.4
 
 ### Patch Changes