+++ packages/installer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pact-toolbox/installer
 
+## 0.0.4
+
+### Patch Changes
+
+- more pact version manager features
+- Updated dependencies
+  - @pact-toolbox/utils@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes